refactor: reorganizar estrutura de pacotes
