+++ src/main/java/com/smartlist/api/user/dto/RegisterDTO.java
@@ -1,4 +1,6 @@
-package com.smartlist.api.user;
+package com.smartlist.api.user.dto;
+
+import com.smartlist.api.user.enums.NotificationPreference;
 
 public record RegisterDTO(String email, String password, String phoneNumber, NotificationPreference notificationPreference) {
 }
