+++ src/main/java/com/smartlist/api/user/repository/UserRepository.java
@@ -1,5 +1,6 @@
-package com.smartlist.api.user;
+package com.smartlist.api.user.repository;
 
+import com.smartlist.api.user.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
